Random Effect Model (REM) MetaVolcano

This function performs Random Effect Model meta-analysis on differential expression results. This function runs the 'Random Effect Model' MetaVolcano section

rem_mv(
  diffexp = list(),
  pcriteria = "pvalue",
  foldchangecol = "Log2FC",
  genenamecol = "Symbol",
  geneidcol = NULL,
  collaps = FALSE,
  llcol = "CI.L",
  rlcol = "CI.R",
  vcol = NULL,
  cvar = TRUE,
  metathr = 0.01,
  jobname = "MetaVolcano",
  outputfolder = tempdir(),
  draw = "HTML",
  ncores = 1,
  colors = c(low = "#083e46", mid = "white", high = "#811820", na = "grey80"),
  point_size = 0.6,
  label_genes = NULL,
  label_top_n = NULL,
  label_size = 3,
  plot_title = NULL,
  show_legend = TRUE
  outputfolder = ".",
  draw = "HTML",
  ncores = 1,
  render = F
)

Arguments

diffexp

list of data.frame/data.table (s) with DE results where lines are genes

pcriteria

the column name of the pvalue variable <string>

foldchangecol

the column name of the foldchange variable <string>

genenamecol

the column name of the gene name variable <string>

geneidcol

the column name of the gene ID/probe/oligo/transcript variable <string>

collaps

if probes should be collapsed based on the DE direction <logical>

llcol

left limit of the fold change coinfidence interval variable name <string>

rlcol

right limit of the fold change coinfidence interval variable name <string>

vcol

name of the fold change variance variable <string>

cvar

weather or not to calculate gene variance from confidence interval limits <logical>

metathr

top percentage of perturbed genes to be highlighted <double>

jobname

name of the running job <string>

outputfolder

/path where to write the results/

draw

wheather or not to draw the .html visualization <logical>

ncores

the number of processors the user wants to use <integer>

colors

named vector of colors for gradient: c(low, mid, high, na)

point_size

size of points in the plot

label_genes

character vector of specific genes to label

label_top_n

number of top genes (by p-value) to label automatically

label_size

size of gene name labels

plot_title

custom plot title (NULL for no title)

show_legend

whether to show the legend (TRUE/FALSE)

ncores

the number of processors the user wants to use <integer> #' @param render A boolean parameter that determines whether the plot should be rendered. If `TRUE`, the function will produce and save the plot based on the specified `draw` parameter (either as an HTML or PDF file). If `FALSE` (default), no plot will be rendered or saved. It's useful for cases where you might want to run the function for its side effects or calculations without necessarily visualizing the result.

Value

MetaVolcano object

Examples

data(diffexplist)
diffexplist <- lapply(diffexplist, function(del) {
    dplyr::filter(del, grepl("MP", Symbol))
})