A function to draw the 'Vote-counting meta-analysis' MetaVolcano

This function draws the vote-counting meta-analysis MetaVolcano

votecount_mv(
  diffexp = list(),
  pcriteria = "pvalue",
  foldchangecol = "Log2FC",
  genenamecol = "Symbol",
  geneidcol = NULL,
  pvalue = 0.05,
  foldchange = 0,
  metathr = 0.01,
  collaps = FALSE,
  jobname = "MetaVolcano",
  outputfolder = tempdir(),
  draw = "HTML",
  colors = c("#083e46", "grey", "#811820"),
  point_size = 0.5,
  label_genes = NULL,
  label_top_n = NULL,
  label_size = 3,
  plot_title = NULL,
  show_legend = FALSE
)

Arguments

diffexp

list of data.frame/data.table (s) with DE results where lines are genes

pcriteria

the column name of the Pval criteria to consider <string>

foldchangecol

the column name of the foldchange variable <string>

genenamecol

the column name of the gene name variable <string>

geneidcol

the column name of the gene ID/probe/oligo/transcript variable <string>

pvalue

the Pval to use as threshold c(0:1) <double>

foldchange

the foldchange to use as DE threshold c(-Inf: Inf) <double>

metathr

the proportion of studies a gene has to be DEG to be considered cDEG <double>

collaps

if probes should be collapsed based on the DE direction <logical>

jobname

name of the running job <string>

outputfolder

/path where to write the results/

draw

wheather or not to draw a .pdf or .html visualization <c(NULL, 'PDF', 'HTML')>

colors

vector of colors for the plot c(down, neutral, up)

point_size

size of points in the plot

label_genes

character vector of specific genes to label

label_top_n

number of top genes to label automatically

label_size

size of gene name labels

plot_title

custom plot title

show_legend

whether to show the legend

Value

MetaVolcano object

Examples

data(diffexplist)
mv <- votecount_mv(diffexplist)
str(mv)
#> Formal class 'MetaVolcano' [package "MetaVolcanoR"] with 5 slots
#>   ..@ input      :'data.frame':	7894 obs. of  11 variables:
#>   .. ..$ Symbol  : chr [1:7894] "ABCC3" "ABHD5" "ACACB" "ACVR1C" ...
#>   .. ..$ Log2FC_1: num [1:7894] 2.788 -1.32 -0.843 -1.632 0.631 ...
#>   .. ..$ pvalue_1: num [1:7894] 1.86e-07 2.63e-07 1.12e-02 5.67e-09 2.01e-02 ...
#>   .. ..$ Log2FC_2: num [1:7894] 1.382 -1.29 -1.112 -1.471 0.858 ...
#>   .. ..$ pvalue_2: num [1:7894] 1.54e-03 1.62e-04 1.31e-03 5.22e-05 7.22e-03 ...
#>   .. ..$ Log2FC_3: num [1:7894] 0.525 -0.438 -0.531 -0.677 0.264 ...
#>   .. ..$ pvalue_3: num [1:7894] 1.41e-06 1.19e-07 3.13e-08 2.11e-09 1.58e-04 ...
#>   .. ..$ Log2FC_4: num [1:7894] 0.488 -0.866 -0.799 -1.984 0.609 ...
#>   .. ..$ pvalue_4: num [1:7894] 0.035539 0.000941 0.020201 0.005407 0.008965 ...
#>   .. ..$ Log2FC_5: num [1:7894] 0.373 -0.795 -0.516 -0.508 0.366 ...
#>   .. ..$ pvalue_5: num [1:7894] 0.04338 0.00232 0.00573 0.02088 0.0366 ...
#>   ..@ inputnames : chr [1:5] "GSE12050" "GSE24883" "GSE25401" "GSE27949" ...
#>   ..@ metaresult :'data.frame':	7894 obs. of  10 variables:
#>   .. ..$ Symbol   : chr [1:7894] "ABCC3" "ABHD5" "ACACB" "ACVR1C" ...
#>   .. ..$ deg_1    : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>   .. ..$ deg_2    : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>   .. ..$ deg_3    : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>   .. ..$ deg_4    : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>   .. ..$ deg_5    : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>   .. ..$ ndeg     : num [1:7894] 5 5 5 5 5 5 5 5 5 5 ...
#>   .. ..$ ddeg     : num [1:7894] 5 -5 -5 -5 5 -5 -5 -5 -5 -5 ...
#>   .. ..$ idx      : num [1:7894] 25 -25 -25 -25 25 -25 -25 -25 -25 -25 ...
#>   .. ..$ degvcount: chr [1:7894] "2.Up-regulated" "0.Down-regulated" "0.Down-regulated" "0.Down-regulated" ...
#>   ..@ MetaVolcano: <ggplot2::ggplot>
#>   .. ..@ data       :'data.frame':	7894 obs. of  21 variables:
#>  .. .. .. $ Symbol    : chr [1:7894] "ABCC3" "ABHD5" "ACACB" "ACVR1C" ...
#>  .. .. .. $ Log2FC_1  : num [1:7894] 2.788 -1.32 -0.843 -1.632 0.631 ...
#>  .. .. .. $ pvalue_1  : num [1:7894] 1.86e-07 2.63e-07 1.12e-02 5.67e-09 2.01e-02 ...
#>  .. .. .. $ deg_1     : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>  .. .. .. $ Log2FC_2  : num [1:7894] 1.382 -1.29 -1.112 -1.471 0.858 ...
#>  .. .. .. $ pvalue_2  : num [1:7894] 1.54e-03 1.62e-04 1.31e-03 5.22e-05 7.22e-03 ...
#>  .. .. .. $ deg_2     : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>  .. .. .. $ Log2FC_3  : num [1:7894] 0.525 -0.438 -0.531 -0.677 0.264 ...
#>  .. .. .. $ pvalue_3  : num [1:7894] 1.41e-06 1.19e-07 3.13e-08 2.11e-09 1.58e-04 ...
#>  .. .. .. $ deg_3     : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>  .. .. .. $ Log2FC_4  : num [1:7894] 0.488 -0.866 -0.799 -1.984 0.609 ...
#>  .. .. .. $ pvalue_4  : num [1:7894] 0.035539 0.000941 0.020201 0.005407 0.008965 ...
#>  .. .. .. $ deg_4     : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>  .. .. .. $ Log2FC_5  : num [1:7894] 0.373 -0.795 -0.516 -0.508 0.366 ...
#>  .. .. .. $ pvalue_5  : num [1:7894] 0.04338 0.00232 0.00573 0.02088 0.0366 ...
#>  .. .. .. $ deg_5     : num [1:7894] 1 -1 -1 -1 1 -1 -1 -1 -1 -1 ...
#>  .. .. .. $ ndeg      : num [1:7894] 5 5 5 5 5 5 5 5 5 5 ...
#>  .. .. .. $ ddeg      : num [1:7894] 5 -5 -5 -5 5 -5 -5 -5 -5 -5 ...
#>  .. .. .. $ idx       : num [1:7894] 25 -25 -25 -25 25 -25 -25 -25 -25 -25 ...
#>  .. .. .. $ degvcount : chr [1:7894] "2.Up-regulated" "0.Down-regulated" "0.Down-regulated" "0.Down-regulated" ...
#>  .. .. .. $ gene_label: chr [1:7894] "" "" "" "" ...
#>   .. ..@ layers     :List of 1
#>  .. .. .. $ geom_jitter:Classes 'LayerInstance', 'Layer', 'ggproto', 'gg' <ggproto object: Class LayerInstance, Layer, gg>
#>     aes_params: list
#>     compute_aesthetics: function
#>     compute_geom_1: function
#>     compute_geom_2: function
#>     compute_position: function
#>     compute_statistic: function
#>     computed_geom_params: list
#>     computed_mapping: ggplot2::mapping, uneval, gg, S7_object
#>     computed_stat_params: list
#>     constructor: call
#>     data: waiver
#>     draw_geom: function
#>     finish_statistics: function
#>     geom: <ggproto object: Class GeomPoint, Geom, gg>
#>         aesthetics: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         draw_group: function
#>         draw_key: function
#>         draw_layer: function
#>         draw_panel: function
#>         extra_params: na.rm
#>         handle_na: function
#>         non_missing_aes: size shape colour
#>         optional_aes: 
#>         parameters: function
#>         rename_size: FALSE
#>         required_aes: x y
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class Geom, gg>
#>     geom_params: list
#>     inherit.aes: TRUE
#>     layer_data: function
#>     layout: NULL
#>     map_statistic: function
#>     mapping: ggplot2::mapping, uneval, gg, S7_object
#>     name: NULL
#>     position: <ggproto object: Class PositionJitter, Position, gg>
#>         aesthetics: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         height: 0.45
#>         required_aes: x y
#>         seed: NA
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         width: 0.45
#>         super:  <ggproto object: Class PositionJitter, Position, gg>
#>     print: function
#>     setup_layer: function
#>     show.legend: NA
#>     stat: <ggproto object: Class StatIdentity, Stat, gg>
#>         aesthetics: function
#>         compute_group: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         dropped_aes: x_plotlyDomain y_plotlyDomain x_plotlyDomain y_plotlyDom ...
#>         extra_params: na.rm
#>         finish_layer: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         required_aes: 
#>         retransform: TRUE
#>         setup_data: function
#>         setup_params: function
#>         super:  <ggproto object: Class Stat, gg>
#>     stat_params: list
#>     super:  <ggproto object: Class Layer, gg> 
#>   .. ..@ scales     :Classes 'ScalesList', 'ggproto', 'gg' <ggproto object: Class ScalesList, gg>
#>     add: function
#>     add_defaults: function
#>     add_missing: function
#>     backtransform_df: function
#>     clone: function
#>     find: function
#>     get_scales: function
#>     has_scale: function
#>     input: function
#>     map_df: function
#>     n: function
#>     non_position_scales: function
#>     scales: list
#>     set_palettes: function
#>     train_df: function
#>     transform_df: function
#>     super:  <ggproto object: Class ScalesList, gg> 
#>   .. ..@ guides     :Classes 'Guides', 'ggproto', 'gg' <ggproto object: Class Guides, gg>
#>     add: function
#>     assemble: function
#>     build: function
#>     draw: function
#>     get_custom: function
#>     get_guide: function
#>     get_params: function
#>     get_position: function
#>     guides: NULL
#>     merge: function
#>     missing: <ggproto object: Class GuideNone, Guide, gg>
#>         add_title: function
#>         arrange_layout: function
#>         assemble_drawing: function
#>         available_aes: any
#>         build_decor: function
#>         build_labels: function
#>         build_ticks: function
#>         build_title: function
#>         draw: function
#>         draw_early_exit: function
#>         elements: list
#>         extract_decor: function
#>         extract_key: function
#>         extract_params: function
#>         get_layer_key: function
#>         hashables: list
#>         measure_grobs: function
#>         merge: function
#>         override_elements: function
#>         params: list
#>         process_layers: function
#>         setup_elements: function
#>         setup_params: function
#>         train: function
#>         transform: function
#>         super:  <ggproto object: Class GuideNone, Guide, gg>
#>     package_box: function
#>     print: function
#>     process_layers: function
#>     setup: function
#>     subset_guides: function
#>     train: function
#>     update_params: function
#>     super:  <ggproto object: Class Guides, gg> 
#>   .. ..@ mapping    : <ggplot2::mapping> List of 3
#>  .. .. .. $ x   : language ~ddeg
#>  .. .. ..  ..- attr(*, ".Environment")=<environment: 0x00000114b4cb36c8> 
#>  .. .. .. $ y   : language ~ndeg
#>  .. .. ..  ..- attr(*, ".Environment")=<environment: 0x00000114b4cb36c8> 
#>  .. .. .. $ text: language ~Symbol
#>  .. .. ..  ..- attr(*, ".Environment")=<environment: 0x00000114b4cb36c8> 
#>   .. ..@ theme      : <theme> List of 144
#>  .. .. .. $ line                            : <ggplot2::element_line>
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ linewidth    : num 0.5
#>  .. .. ..  ..@ linetype     : num 1
#>  .. .. ..  ..@ lineend      : chr "butt"
#>  .. .. ..  ..@ linejoin     : chr "round"
#>  .. .. ..  ..@ arrow        : logi FALSE
#>  .. .. ..  ..@ arrow.fill   : chr "black"
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ rect                            : <ggplot2::element_rect>
#>  .. .. ..  ..@ fill         : chr "white"
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ linewidth    : num 0.5
#>  .. .. ..  ..@ linetype     : num 1
#>  .. .. ..  ..@ linejoin     : chr "round"
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ text                            : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : chr ""
#>  .. .. ..  ..@ face         : chr "plain"
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ size         : num 11
#>  .. .. ..  ..@ hjust        : num 0.5
#>  .. .. ..  ..@ vjust        : num 0.5
#>  .. .. ..  ..@ angle        : num 0
#>  .. .. ..  ..@ lineheight   : num 0.9
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 0
#>  .. .. ..  ..@ debug        : logi FALSE
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ title                           : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ point                           : <ggplot2::element_point>
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ shape        : num 19
#>  .. .. ..  ..@ size         : num 1.5
#>  .. .. ..  ..@ fill         : chr "white"
#>  .. .. ..  ..@ stroke       : num 0.5
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ polygon                         : <ggplot2::element_polygon>
#>  .. .. ..  ..@ fill         : chr "white"
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ linewidth    : num 0.5
#>  .. .. ..  ..@ linetype     : num 1
#>  .. .. ..  ..@ linejoin     : chr "round"
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ geom                            : <ggplot2::element_geom>
#>  .. .. ..  ..@ ink        : chr "black"
#>  .. .. ..  ..@ paper      : chr "white"
#>  .. .. ..  ..@ accent     : chr "#3366FF"
#>  .. .. ..  ..@ linewidth  : num 0.5
#>  .. .. ..  ..@ borderwidth: num 0.5
#>  .. .. ..  ..@ linetype   : int 1
#>  .. .. ..  ..@ bordertype : int 1
#>  .. .. ..  ..@ family     : chr ""
#>  .. .. ..  ..@ fontsize   : num 3.87
#>  .. .. ..  ..@ pointsize  : num 1.5
#>  .. .. ..  ..@ pointshape : num 19
#>  .. .. ..  ..@ colour     : NULL
#>  .. .. ..  ..@ fill       : NULL
#>  .. .. .. $ spacing                         : 'simpleUnit' num 5.5points
#>  .. .. ..  ..- attr(*, "unit")= int 8
#>  .. .. .. $ margins                         : <ggplot2::margin> num [1:4] 5.5 5.5 5.5 5.5
#>  .. .. .. $ aspect.ratio                    : NULL
#>  .. .. .. $ axis.title                      : NULL
#>  .. .. .. $ axis.title.x                    : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 1
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 2.75 0 0 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.title.x.top                : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 0
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 2.75 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.title.x.bottom             : NULL
#>  .. .. .. $ axis.title.y                    : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 1
#>  .. .. ..  ..@ angle        : num 90
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.75 0 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.title.y.left               : NULL
#>  .. .. .. $ axis.title.y.right              : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 1
#>  .. .. ..  ..@ angle        : num -90
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 2.75
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.text                       : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : 'rel' num 0.8
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.text.x                     : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 0.5
#>  .. .. ..  ..@ angle        : num 0
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 2.2 0 0 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. .. $ axis.text.x.top                 : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : num 0
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 2.2 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.text.x.bottom              : NULL
#>  .. .. .. $ axis.text.y                     : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : num 1
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.2 0 0
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.text.y.left                : NULL
#>  .. .. .. $ axis.text.y.right               : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : num 0
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 2.2
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.text.theta                 : NULL
#>  .. .. .. $ axis.text.r                     : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : num 0.5
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.2 0 2.2
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.ticks                      : <ggplot2::element_line>
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ linewidth    : NULL
#>  .. .. ..  ..@ linetype     : NULL
#>  .. .. ..  ..@ lineend      : NULL
#>  .. .. ..  ..@ linejoin     : NULL
#>  .. .. ..  ..@ arrow        : logi FALSE
#>  .. .. ..  ..@ arrow.fill   : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.ticks.x                    : NULL
#>  .. .. .. $ axis.ticks.x.top                : NULL
#>  .. .. .. $ axis.ticks.x.bottom             : NULL
#>  .. .. .. $ axis.ticks.y                    : NULL
#>  .. .. .. $ axis.ticks.y.left               : NULL
#>  .. .. .. $ axis.ticks.y.right              : NULL
#>  .. .. .. $ axis.ticks.theta                : NULL
#>  .. .. .. $ axis.ticks.r                    : NULL
#>  .. .. .. $ axis.minor.ticks.x.top          : NULL
#>  .. .. .. $ axis.minor.ticks.x.bottom       : NULL
#>  .. .. .. $ axis.minor.ticks.y.left         : NULL
#>  .. .. .. $ axis.minor.ticks.y.right        : NULL
#>  .. .. .. $ axis.minor.ticks.theta          : NULL
#>  .. .. .. $ axis.minor.ticks.r              : NULL
#>  .. .. .. $ axis.ticks.length               : 'rel' num 0.5
#>  .. .. .. $ axis.ticks.length.x             : NULL
#>  .. .. .. $ axis.ticks.length.x.top         : NULL
#>  .. .. .. $ axis.ticks.length.x.bottom      : NULL
#>  .. .. .. $ axis.ticks.length.y             : NULL
#>  .. .. .. $ axis.ticks.length.y.left        : NULL
#>  .. .. .. $ axis.ticks.length.y.right       : NULL
#>  .. .. .. $ axis.ticks.length.theta         : NULL
#>  .. .. .. $ axis.ticks.length.r             : NULL
#>  .. .. .. $ axis.minor.ticks.length         : 'rel' num 0.75
#>  .. .. .. $ axis.minor.ticks.length.x       : NULL
#>  .. .. .. $ axis.minor.ticks.length.x.top   : NULL
#>  .. .. .. $ axis.minor.ticks.length.x.bottom: NULL
#>  .. .. .. $ axis.minor.ticks.length.y       : NULL
#>  .. .. .. $ axis.minor.ticks.length.y.left  : NULL
#>  .. .. .. $ axis.minor.ticks.length.y.right : NULL
#>  .. .. .. $ axis.minor.ticks.length.theta   : NULL
#>  .. .. .. $ axis.minor.ticks.length.r       : NULL
#>  .. .. .. $ axis.line                       : <ggplot2::element_line>
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ linewidth    : NULL
#>  .. .. ..  ..@ linetype     : NULL
#>  .. .. ..  ..@ lineend      : chr "square"
#>  .. .. ..  ..@ linejoin     : NULL
#>  .. .. ..  ..@ arrow        : logi FALSE
#>  .. .. ..  ..@ arrow.fill   : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ axis.line.x                     : <ggplot2::element_line>
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ linewidth    : num 0.6
#>  .. .. ..  ..@ linetype     : NULL
#>  .. .. ..  ..@ lineend      : chr "square"
#>  .. .. ..  ..@ linejoin     : NULL
#>  .. .. ..  ..@ arrow        : logi FALSE
#>  .. .. ..  ..@ arrow.fill   : chr "black"
#>  .. .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. .. $ axis.line.x.top                 : NULL
#>  .. .. .. $ axis.line.x.bottom              : NULL
#>  .. .. .. $ axis.line.y                     : <ggplot2::element_line>
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ linewidth    : num 0.6
#>  .. .. ..  ..@ linetype     : NULL
#>  .. .. ..  ..@ lineend      : chr "square"
#>  .. .. ..  ..@ linejoin     : NULL
#>  .. .. ..  ..@ arrow        : logi FALSE
#>  .. .. ..  ..@ arrow.fill   : chr "black"
#>  .. .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. .. $ axis.line.y.left                : NULL
#>  .. .. .. $ axis.line.y.right               : NULL
#>  .. .. .. $ axis.line.theta                 : NULL
#>  .. .. .. $ axis.line.r                     : NULL
#>  .. .. .. $ legend.background               : <ggplot2::element_rect>
#>  .. .. ..  ..@ fill         : NULL
#>  .. .. ..  ..@ colour       : logi NA
#>  .. .. ..  ..@ linewidth    : NULL
#>  .. .. ..  ..@ linetype     : NULL
#>  .. .. ..  ..@ linejoin     : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ legend.margin                   : NULL
#>  .. .. .. $ legend.spacing                  : 'rel' num 2
#>  .. .. .. $ legend.spacing.x                : NULL
#>  .. .. .. $ legend.spacing.y                : NULL
#>  .. .. .. $ legend.key                      : NULL
#>  .. .. .. $ legend.key.size                 : 'simpleUnit' num 1.2lines
#>  .. .. ..  ..- attr(*, "unit")= int 3
#>  .. .. .. $ legend.key.height               : NULL
#>  .. .. .. $ legend.key.width                : NULL
#>  .. .. .. $ legend.key.spacing              : NULL
#>  .. .. .. $ legend.key.spacing.x            : NULL
#>  .. .. .. $ legend.key.spacing.y            : NULL
#>  .. .. .. $ legend.key.justification        : NULL
#>  .. .. .. $ legend.frame                    : NULL
#>  .. .. .. $ legend.ticks                    : NULL
#>  .. .. .. $ legend.ticks.length             : 'rel' num 0.2
#>  .. .. .. $ legend.axis.line                : NULL
#>  .. .. .. $ legend.text                     : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : 'rel' num 0.8
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ legend.text.position            : NULL
#>  .. .. .. $ legend.title                    : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : num 0
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ legend.title.position           : NULL
#>  .. .. .. $ legend.position                 : chr "none"
#>  .. .. .. $ legend.position.inside          : NULL
#>  .. .. .. $ legend.direction                : NULL
#>  .. .. .. $ legend.byrow                    : NULL
#>  .. .. .. $ legend.justification            : chr "center"
#>  .. .. .. $ legend.justification.top        : NULL
#>  .. .. .. $ legend.justification.bottom     : NULL
#>  .. .. .. $ legend.justification.left       : NULL
#>  .. .. .. $ legend.justification.right      : NULL
#>  .. .. .. $ legend.justification.inside     : NULL
#>  .. .. ..  [list output truncated]
#>  .. .. .. @ complete: logi TRUE
#>  .. .. .. @ validate: logi TRUE
#>   .. ..@ coordinates:Classes 'CoordCartesian', 'Coord', 'ggproto', 'gg' <ggproto object: Class CoordCartesian, Coord, gg>
#>     aspect: function
#>     backtransform_range: function
#>     clip: on
#>     default: TRUE
#>     distance: function
#>     draw_panel: function
#>     expand: TRUE
#>     is_free: function
#>     is_linear: function
#>     labels: function
#>     limits: list
#>     modify_scales: function
#>     range: function
#>     ratio: NULL
#>     render_axis_h: function
#>     render_axis_v: function
#>     render_bg: function
#>     render_fg: function
#>     reverse: none
#>     setup_data: function
#>     setup_layout: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     setup_params: function
#>     train_panel_guides: function
#>     transform: function
#>     super:  <ggproto object: Class CoordCartesian, Coord, gg> 
#>   .. ..@ facet      :Classes 'FacetNull', 'Facet', 'ggproto', 'gg' <ggproto object: Class FacetNull, Facet, gg>
#>     attach_axes: function
#>     attach_strips: function
#>     compute_layout: function
#>     draw_back: function
#>     draw_front: function
#>     draw_labels: function
#>     draw_panel_content: function
#>     draw_panels: function
#>     finish_data: function
#>     format_strip_labels: function
#>     init_gtable: function
#>     init_scales: function
#>     map_data: function
#>     params: list
#>     set_panel_size: function
#>     setup_data: function
#>     setup_panel_params: function
#>     setup_params: function
#>     shrink: TRUE
#>     train_scales: function
#>     vars: function
#>     super:  <ggproto object: Class FacetNull, Facet, gg> 
#>   .. ..@ layout     :Classes 'Layout', 'ggproto', 'gg' <ggproto object: Class Layout, gg>
#>     coord: NULL
#>     coord_params: list
#>     facet: NULL
#>     facet_params: list
#>     finish_data: function
#>     get_scales: function
#>     layout: NULL
#>     map_position: function
#>     panel_params: NULL
#>     panel_scales_x: NULL
#>     panel_scales_y: NULL
#>     render: function
#>     render_labels: function
#>     reset_scales: function
#>     resolve_label: function
#>     setup: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     train_position: function
#>     super:  <ggproto object: Class Layout, gg> 
#>   .. ..@ labels     : <ggplot2::labels> List of 3
#>  .. .. .. $ x    : chr "Sign consistency"
#>  .. .. .. $ y    : chr "Differential feature frequency"
#>  .. .. .. $ title: NULL
#>   .. ..@ meta       : list()
#>   .. ..@ plot_env   :<environment: 0x00000114b4cb36c8> 
#>   ..@ featurefreq: <ggplot2::ggplot>
#>   .. ..@ data       : list()
#>  .. .. .. - attr(*, "class")= chr "waiver"
#>   .. ..@ layers     :List of 1
#>  .. .. .. $ draw_grob:Classes 'LayerInstance', 'Layer', 'ggproto', 'gg' <ggproto object: Class LayerInstance, Layer, gg>
#>     aes_params: list
#>     compute_aesthetics: function
#>     compute_geom_1: function
#>     compute_geom_2: function
#>     compute_position: function
#>     compute_statistic: function
#>     computed_geom_params: NULL
#>     computed_mapping: NULL
#>     computed_stat_params: NULL
#>     constructor: call
#>     data: data.frame
#>     draw_geom: function
#>     finish_statistics: function
#>     geom: <ggproto object: Class GeomDrawGrob, GeomCustomAnn, Geom, gg>
#>         aesthetics: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         draw_group: function
#>         draw_key: function
#>         draw_layer: function
#>         draw_panel: function
#>         extra_params: 
#>         handle_na: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         rename_size: FALSE
#>         required_aes: 
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class GeomCustomAnn, Geom, gg>
#>     geom_params: list
#>     inherit.aes: FALSE
#>     layer_data: function
#>     layout: NULL
#>     map_statistic: function
#>     mapping: NULL
#>     name: NULL
#>     position: <ggproto object: Class PositionIdentity, Position, gg>
#>         aesthetics: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         required_aes: 
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class Position, gg>
#>     print: function
#>     setup_layer: function
#>     show.legend: NA
#>     stat: <ggproto object: Class StatIdentity, Stat, gg>
#>         aesthetics: function
#>         compute_group: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         dropped_aes: x_plotlyDomain y_plotlyDomain x_plotlyDomain y_plotlyDom ...
#>         extra_params: na.rm
#>         finish_layer: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         required_aes: 
#>         retransform: TRUE
#>         setup_data: function
#>         setup_params: function
#>         super:  <ggproto object: Class Stat, gg>
#>     stat_params: list
#>     super:  <ggproto object: Class Layer, gg> 
#>   .. ..@ scales     :Classes 'ScalesList', 'ggproto', 'gg' <ggproto object: Class ScalesList, gg>
#>     add: function
#>     add_defaults: function
#>     add_missing: function
#>     backtransform_df: function
#>     clone: function
#>     find: function
#>     get_scales: function
#>     has_scale: function
#>     input: function
#>     map_df: function
#>     n: function
#>     non_position_scales: function
#>     scales: list
#>     set_palettes: function
#>     train_df: function
#>     transform_df: function
#>     super:  <ggproto object: Class ScalesList, gg> 
#>   .. ..@ guides     :Classes 'Guides', 'ggproto', 'gg' <ggproto object: Class Guides, gg>
#>     add: function
#>     assemble: function
#>     build: function
#>     draw: function
#>     get_custom: function
#>     get_guide: function
#>     get_params: function
#>     get_position: function
#>     guides: NULL
#>     merge: function
#>     missing: <ggproto object: Class GuideNone, Guide, gg>
#>         add_title: function
#>         arrange_layout: function
#>         assemble_drawing: function
#>         available_aes: any
#>         build_decor: function
#>         build_labels: function
#>         build_ticks: function
#>         build_title: function
#>         draw: function
#>         draw_early_exit: function
#>         elements: list
#>         extract_decor: function
#>         extract_key: function
#>         extract_params: function
#>         get_layer_key: function
#>         hashables: list
#>         measure_grobs: function
#>         merge: function
#>         override_elements: function
#>         params: list
#>         process_layers: function
#>         setup_elements: function
#>         setup_params: function
#>         train: function
#>         transform: function
#>         super:  <ggproto object: Class GuideNone, Guide, gg>
#>     package_box: function
#>     print: function
#>     process_layers: function
#>     setup: function
#>     subset_guides: function
#>     train: function
#>     update_params: function
#>     super:  <ggproto object: Class Guides, gg> 
#>   .. ..@ mapping    : <ggplot2::mapping>  Named list()
#>   .. ..@ theme      : <theme> List of 144
#>  .. .. .. $ line                            : <ggplot2::element_blank>
#>  .. .. .. $ rect                            : <ggplot2::element_blank>
#>  .. .. .. $ text                            : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : chr ""
#>  .. .. ..  ..@ face         : chr "plain"
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : chr "black"
#>  .. .. ..  ..@ size         : num 14
#>  .. .. ..  ..@ hjust        : num 0.5
#>  .. .. ..  ..@ vjust        : num 0.5
#>  .. .. ..  ..@ angle        : num 0
#>  .. .. ..  ..@ lineheight   : num 0.9
#>  .. .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 0
#>  .. .. ..  ..@ debug        : logi FALSE
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ title                           : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ point                           : <ggplot2::element_blank>
#>  .. .. .. $ polygon                         : <ggplot2::element_blank>
#>  .. .. .. $ geom                            : <ggplot2::element_geom>
#>  .. .. ..  ..@ ink        : chr "black"
#>  .. .. ..  ..@ paper      : chr "#00000000"
#>  .. .. ..  ..@ accent     : chr "#3366FF"
#>  .. .. ..  ..@ linewidth  : num 0.636
#>  .. .. ..  ..@ borderwidth: num 0.636
#>  .. .. ..  ..@ linetype   : int 1
#>  .. .. ..  ..@ bordertype : int 1
#>  .. .. ..  ..@ family     : chr ""
#>  .. .. ..  ..@ fontsize   : num 4.92
#>  .. .. ..  ..@ pointsize  : num 1.91
#>  .. .. ..  ..@ pointshape : num 19
#>  .. .. ..  ..@ colour     : NULL
#>  .. .. ..  ..@ fill       : NULL
#>  .. .. .. $ spacing                         : 'simpleUnit' num 7points
#>  .. .. ..  ..- attr(*, "unit")= int 8
#>  .. .. .. $ margins                         : <ggplot2::margin> num [1:4] 7 7 7 7
#>  .. .. .. $ aspect.ratio                    : NULL
#>  .. .. .. $ axis.title                      : <ggplot2::element_blank>
#>  .. .. .. $ axis.title.x                    : NULL
#>  .. .. .. $ axis.title.x.top                : NULL
#>  .. .. .. $ axis.title.x.bottom             : NULL
#>  .. .. .. $ axis.title.y                    : NULL
#>  .. .. .. $ axis.title.y.left               : NULL
#>  .. .. .. $ axis.title.y.right              : NULL
#>  .. .. .. $ axis.text                       : <ggplot2::element_blank>
#>  .. .. .. $ axis.text.x                     : NULL
#>  .. .. .. $ axis.text.x.top                 : NULL
#>  .. .. .. $ axis.text.x.bottom              : NULL
#>  .. .. .. $ axis.text.y                     : NULL
#>  .. .. .. $ axis.text.y.left                : NULL
#>  .. .. .. $ axis.text.y.right               : NULL
#>  .. .. .. $ axis.text.theta                 : NULL
#>  .. .. .. $ axis.text.r                     : NULL
#>  .. .. .. $ axis.ticks                      : <ggplot2::element_blank>
#>  .. .. .. $ axis.ticks.x                    : NULL
#>  .. .. .. $ axis.ticks.x.top                : NULL
#>  .. .. .. $ axis.ticks.x.bottom             : NULL
#>  .. .. .. $ axis.ticks.y                    : NULL
#>  .. .. .. $ axis.ticks.y.left               : NULL
#>  .. .. .. $ axis.ticks.y.right              : NULL
#>  .. .. .. $ axis.ticks.theta                : NULL
#>  .. .. .. $ axis.ticks.r                    : NULL
#>  .. .. .. $ axis.minor.ticks.x.top          : NULL
#>  .. .. .. $ axis.minor.ticks.x.bottom       : NULL
#>  .. .. .. $ axis.minor.ticks.y.left         : NULL
#>  .. .. .. $ axis.minor.ticks.y.right        : NULL
#>  .. .. .. $ axis.minor.ticks.theta          : NULL
#>  .. .. .. $ axis.minor.ticks.r              : NULL
#>  .. .. .. $ axis.ticks.length               : 'simpleUnit' num 0points
#>  .. .. ..  ..- attr(*, "unit")= int 8
#>  .. .. .. $ axis.ticks.length.x             : NULL
#>  .. .. .. $ axis.ticks.length.x.top         : NULL
#>  .. .. .. $ axis.ticks.length.x.bottom      : NULL
#>  .. .. .. $ axis.ticks.length.y             : NULL
#>  .. .. .. $ axis.ticks.length.y.left        : NULL
#>  .. .. .. $ axis.ticks.length.y.right       : NULL
#>  .. .. .. $ axis.ticks.length.theta         : NULL
#>  .. .. .. $ axis.ticks.length.r             : NULL
#>  .. .. .. $ axis.minor.ticks.length         : NULL
#>  .. .. .. $ axis.minor.ticks.length.x       : NULL
#>  .. .. .. $ axis.minor.ticks.length.x.top   : NULL
#>  .. .. .. $ axis.minor.ticks.length.x.bottom: NULL
#>  .. .. .. $ axis.minor.ticks.length.y       : NULL
#>  .. .. .. $ axis.minor.ticks.length.y.left  : NULL
#>  .. .. .. $ axis.minor.ticks.length.y.right : NULL
#>  .. .. .. $ axis.minor.ticks.length.theta   : NULL
#>  .. .. .. $ axis.minor.ticks.length.r       : NULL
#>  .. .. .. $ axis.line                       : <ggplot2::element_blank>
#>  .. .. .. $ axis.line.x                     : NULL
#>  .. .. .. $ axis.line.x.top                 : NULL
#>  .. .. .. $ axis.line.x.bottom              : NULL
#>  .. .. .. $ axis.line.y                     : NULL
#>  .. .. .. $ axis.line.y.left                : NULL
#>  .. .. .. $ axis.line.y.right               : NULL
#>  .. .. .. $ axis.line.theta                 : NULL
#>  .. .. .. $ axis.line.r                     : NULL
#>  .. .. .. $ legend.background               : <ggplot2::element_blank>
#>  .. .. .. $ legend.margin                   : <ggplot2::margin> num [1:4] 0 0 0 0
#>  .. .. .. $ legend.spacing                  : 'simpleUnit' num 14points
#>  .. .. ..  ..- attr(*, "unit")= int 8
#>  .. .. .. $ legend.spacing.x                : NULL
#>  .. .. .. $ legend.spacing.y                : NULL
#>  .. .. .. $ legend.key                      : <ggplot2::element_blank>
#>  .. .. .. $ legend.key.size                 : 'simpleUnit' num 15.4points
#>  .. .. ..  ..- attr(*, "unit")= int 8
#>  .. .. .. $ legend.key.height               : NULL
#>  .. .. .. $ legend.key.width                : NULL
#>  .. .. .. $ legend.key.spacing              : 'rel' num 1
#>  .. .. .. $ legend.key.spacing.x            : NULL
#>  .. .. .. $ legend.key.spacing.y            : NULL
#>  .. .. .. $ legend.key.justification        : NULL
#>  .. .. .. $ legend.frame                    : <ggplot2::element_blank>
#>  .. .. .. $ legend.ticks                    : NULL
#>  .. .. .. $ legend.ticks.length             : 'rel' num 0.2
#>  .. .. .. $ legend.axis.line                : NULL
#>  .. .. .. $ legend.text                     : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : 'rel' num 0.857
#>  .. .. ..  ..@ hjust        : NULL
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ legend.text.position            : NULL
#>  .. .. .. $ legend.title                    : <ggplot2::element_text>
#>  .. .. ..  ..@ family       : NULL
#>  .. .. ..  ..@ face         : NULL
#>  .. .. ..  ..@ italic       : chr NA
#>  .. .. ..  ..@ fontweight   : num NA
#>  .. .. ..  ..@ fontwidth    : num NA
#>  .. .. ..  ..@ colour       : NULL
#>  .. .. ..  ..@ size         : NULL
#>  .. .. ..  ..@ hjust        : num 0
#>  .. .. ..  ..@ vjust        : NULL
#>  .. .. ..  ..@ angle        : NULL
#>  .. .. ..  ..@ lineheight   : NULL
#>  .. .. ..  ..@ margin       : NULL
#>  .. .. ..  ..@ debug        : NULL
#>  .. .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. .. $ legend.title.position           : NULL
#>  .. .. .. $ legend.position                 : chr "none"
#>  .. .. .. $ legend.position.inside          : NULL
#>  .. .. .. $ legend.direction                : NULL
#>  .. .. .. $ legend.byrow                    : NULL
#>  .. .. .. $ legend.justification            : chr "center"
#>  .. .. .. $ legend.justification.top        : NULL
#>  .. .. .. $ legend.justification.bottom     : NULL
#>  .. .. .. $ legend.justification.left       : NULL
#>  .. .. .. $ legend.justification.right      : NULL
#>  .. .. .. $ legend.justification.inside     : NULL
#>  .. .. ..  [list output truncated]
#>  .. .. .. @ complete: logi TRUE
#>  .. .. .. @ validate: logi TRUE
#>   .. ..@ coordinates:Classes 'CoordCartesian', 'Coord', 'ggproto', 'gg' <ggproto object: Class CoordCartesian, Coord, gg>
#>     aspect: function
#>     backtransform_range: function
#>     clip: off
#>     default: FALSE
#>     distance: function
#>     draw_panel: function
#>     expand: FALSE
#>     is_free: function
#>     is_linear: function
#>     labels: function
#>     limits: list
#>     modify_scales: function
#>     range: function
#>     ratio: NULL
#>     render_axis_h: function
#>     render_axis_v: function
#>     render_bg: function
#>     render_fg: function
#>     reverse: none
#>     setup_data: function
#>     setup_layout: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     setup_params: function
#>     train_panel_guides: function
#>     transform: function
#>     super:  <ggproto object: Class CoordCartesian, Coord, gg> 
#>   .. ..@ facet      :Classes 'FacetNull', 'Facet', 'ggproto', 'gg' <ggproto object: Class FacetNull, Facet, gg>
#>     attach_axes: function
#>     attach_strips: function
#>     compute_layout: function
#>     draw_back: function
#>     draw_front: function
#>     draw_labels: function
#>     draw_panel_content: function
#>     draw_panels: function
#>     finish_data: function
#>     format_strip_labels: function
#>     init_gtable: function
#>     init_scales: function
#>     map_data: function
#>     params: list
#>     set_panel_size: function
#>     setup_data: function
#>     setup_panel_params: function
#>     setup_params: function
#>     shrink: TRUE
#>     train_scales: function
#>     vars: function
#>     super:  <ggproto object: Class FacetNull, Facet, gg> 
#>   .. ..@ layout     :Classes 'Layout', 'ggproto', 'gg' <ggproto object: Class Layout, gg>
#>     coord: NULL
#>     coord_params: list
#>     facet: NULL
#>     facet_params: list
#>     finish_data: function
#>     get_scales: function
#>     layout: NULL
#>     map_position: function
#>     panel_params: NULL
#>     panel_scales_x: NULL
#>     panel_scales_y: NULL
#>     render: function
#>     render_labels: function
#>     reset_scales: function
#>     resolve_label: function
#>     setup: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     train_position: function
#>     super:  <ggproto object: Class Layout, gg> 
#>   .. ..@ labels     : <ggplot2::labels>  Named list()
#>   .. ..@ meta       : list()
#>   .. ..@ plot_env   :<environment: 0x00000114ec2aa708>